refactor:법안상세 의견 리팩토링
