+++ src/features/bill-detail/hooks/useReactionInfo.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { BillReaction } from '../const';
+import { BillReaction, REACTION_ICON_MAP } from '../const';
 import { getMyReactions } from '../api/server';
 
 export const useReactionInfo = (id: string) => {
@@ -17,5 +17,31 @@ export const useReactionInfo = (id: string) => {
 		fetchReactionInfo();
 	}, [id]);
 
-	return { reactionCounts, myReaction, setMyReaction, setReactionCounts };
+	const updateReaction = (index: number) => {
+		const selectedLabel = REACTION_ICON_MAP[index].label;
+		const updated: [number, number, number, number] = [...reactionCounts];
+
+		if (myReaction === selectedLabel) {
+			updated[index] -= 1;
+			setReactionCounts(updated);
+			setMyReaction(null);
+			return;
+		}
+
+		updated[index] += 1;
+
+		if (myReaction !== null) {
+			const prevIndex = REACTION_ICON_MAP.findIndex((item) => item.label === myReaction);
+			if (prevIndex !== -1) {
+				updated[prevIndex] = Math.max(0, updated[prevIndex] - 1);
+			}
+		}
+
+		setReactionCounts(updated);
+		setMyReaction(selectedLabel);
+
+		// TODO: 서버 전송 로직 추가
+	};
+
+	return { reactionCounts, myReaction, setMyReaction, setReactionCounts, updateReaction };
 };
